Added "exec" and a few more aditions
So, today in computer science class, I thought about the possibility of calling files as functions, adding more strength to the "the functions is what matters the most" in DebsLang. It's just a funny addition tho, and I really have to polish it a bit, because it currently generates a lot of garbage. Maybe temporal variables that get destroyed after calling a function file?